Add support for awesome_print calls
Prior to this commit if Awesome_print was used in place of puts we putsinator would return the location of the puts call within the awesome print gem instead of the location where the ap() call was made. This is a simple change so that putsinator is just as useful when api() has been used.